requirements.txt
# Core
python-dotenv==1.0.1
requests
pandas==2.2.3
numpy==2.2.6
matplotlib
plotly==5.18.0
streamlit>=1.32,<2
ta

# FastAPI Backend
fastapi==0.115.12
uvicorn==0.34.2
starlette==0.46.2
pydantic==2.11.5
pydantic_core==2.33.2
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
h11==0.16.0
idna==3.10
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2

# ML & Technical Analysis
scikit-learn
xgboost
tensorflow
keras

# NLP / Sentiment (FinBERT)
transformers
torch
sentencepiece

# Polygon API Client
polygon-api-client==1.14.0

# LangChain / Agents
langchain
langchain-core
langchain-community
openai
tiktoken

# Utils & Logging
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
loguru
aiohttp
httpx